+++ src/main/java/com/example/springforumapp/users/repositories/RolesRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RolesRepository extends JpaRepository<Role, Integer> {
+public interface RolesRepository extends JpaRepository<Role, Long> {
     Role findByName(String name);
 }
